ConformanceTest - Less messy output
This test seems to include some debug output.  I can understand if you guys
want that noisy/debug mode to be available, but for a default run of the test
suite we should just use phpunit's normal output.

This patch is a compromise -- if you want the debug output, then set an
environment variable `API4_DEBUG=1`.