Failing test to demonstrate bug in join logic
When joining 1 phone to 1 contact it appears to work because the database is otherwise empty.
But when there is more than one phone (as is the case in the SingleContact dataset)
it fails to match due to a flaw in the join logic. I believe the bug is somewhere in the
\Civi\Api4\Query\Api4SelectQuery::postRun and ArrayInsertionUtil::insert functions.