Implement BasicBatchAction
Ex: The `Afform.revert` API

Note: In `AbstractBatchAction::__construct`, I renamed `$idField` to
`$select.  Why?  In the `batchFrobnicate()` example (`BasicActionsTest`),
the action uses two pieces of data from each item (i.e.  the `id` and
`number`).

It's true that the specific cases of `update()`/`replace()`/`delete()` only
need an `id` field, but for other tasks generally, it seems reasonable to
pre-fetch a couple pieces of data that may be used in the task.

Additionally, that actually simplifies `AbstractBatchAction`...  since the
field really is a `select` list.